feat: remove concrete dependency, clean up

+++ assemblyscript/assembly/index.ts
@@ -1,7 +1,7 @@
 import { QuicksortBenchmark } from "./sort";
 
-export function run(seed: i64, arrLen: i64, iter: i64): i64 {
+export function run(seed: i32, arrLen: i32, iter: i32): i32 {
   let qs = new QuicksortBenchmark(<u32>seed);
   const checksum = qs.run(<u32>arrLen, <u32>iter);
-  return <i64>checksum;
+  return <i32>checksum;
 }
